requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings>=2.10.1
python-dotenv==1.0.1

# PDF Processing
pypdf==3.17.4
PyMuPDF>=1.23.0  # fitz 모듈
langchain-text-splitters>=0.2.0

# AI/ML - LangChain 1.0 latest
langchain==1.0.3
langchain-openai==1.0.2
langchain-community==0.4.1
langchain-core==1.0.3
openai>=1.109.1,<3.0.0
tiktoken>=0.5.1 

# Vector Store - PostgreSQL with pgvector
pgvector==0.3.6
langchain-postgres==0.0.16

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis (Session & Cache)
redis>=5.0.0

# 스케줄링
apscheduler>=3.10.4

# ML - 감정 분석 모델
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

